cpp: added support for SSE.  The default cpu reported by cpuid will now be a Pentium 3.  To disable SSE you can used the command line argument -p2 to set the cpuid to a Pentium 2.  SSE is not well tested because most of the games that use are Direct3D games which I don't support yet, but Unreal (OpenGL game), uses it for sound and it seems to work.  SSE was implemented across all emulated CPUs (normal core, x86 core and x64 core).

+++ platform/windows/platform.cpp
@@ -114,7 +114,7 @@ int getPixelFormats(PixelFormat* pfd, int maxPfs) {
             pfd[result].dwVisibleMask = p.dwVisibleMask;
             pfd[result].dwDamageMask = p.dwDamageMask;
             result++;
-            fprintf(stderr, "Pixel Format: %d bit (%d%d%d%d) %s:%s depth=%d stencil=%d accum=%d\n", p.cColorBits, p.cRedBits, p.cBlueBits, p.cGreenBits, p.cAlphaBits, (p.dwFlags & K_PFD_GENERIC_FORMAT)?"not accelerated":"accelerated", (p.dwFlags & K_PFD_DOUBLEBUFFER)?"double buffered":"single buffered", p.cDepthBits, p.cStencilBits, p.cAccumBits);
+            fprintf(stderr, "Pixel Format: %d bit (%d%d%d%d) %s:%s depth=%d stencil=%d accum=%d\n", (int)p.cColorBits, (int)p.cRedBits, (int)p.cBlueBits, (int)p.cGreenBits, (int)p.cAlphaBits, (p.dwFlags & K_PFD_GENERIC_FORMAT)?"not accelerated":"accelerated", (p.dwFlags & K_PFD_DOUBLEBUFFER)?"double buffered":"single buffered", (int)p.cDepthBits, (int)p.cStencilBits, (int)p.cAccumBits);
         }
     }
     if (result==1) {
